+++ src/emdbg/debug/jlink.py
@@ -29,21 +29,23 @@ class JLinkBackend(ProbeBackend):
 
     See `call()` for additional information.
     """
-    def __init__(self, device: str, speed: int, rtos: Path = None):
+    def __init__(self, device: str, speed: int, serial: str = None, rtos: Path = None):
         """
         :param device: part name of the microcontroller to debug.
         :param speed: SWD connection baudrate in kHz.
+        :param serial: Serial number of J-Link for selection.
         :param rtos: Path to RTOS plugin for thread-aware debugging.
         """
         super().__init__(":2331")
         self.device = device
         self.speed = speed
+        self.serial = serial
         self.rtos = rtos
         self.process = None
         self.name = "jlink"
 
     def start(self):
-        self.process = call(self.device, self.speed, self.rtos,
+        self.process = call(self.device, self.speed, self.serial, self.rtos,
                             blocking=False, silent=True)
         LOGGER.info(f"Starting {self.process.pid}...")
 
@@ -56,8 +58,8 @@ class JLinkBackend(ProbeBackend):
 
 
 # -----------------------------------------------------------------------------
-def call(device: str, speed: int, rtos: Path = None, blocking: bool = True,
-         silent: bool = False) -> "int | subprocess.Popen":
+def call(device: str, speed: int, serial: str = None, rtos: Path = None,
+         blocking: bool = True, silent: bool = False) -> "int | subprocess.Popen":
     """
     Starts the `JLinkGDBServer` and connects to the microcontroller without
     resetting the device and without GUI. You can overwrite the default binary
@@ -69,6 +71,7 @@ def call(device: str, speed: int, rtos: Path = None, blocking: bool = True,
 
     :param device: part name of the microcontroller to debug.
     :param speed: SWD connection baudrate in kHz.
+    :param serial: Serial number of J-Link for selection.
     :param rtos: Path to RTOS plugin for thread-aware debugging.
     :param blocking: Run in current process as a blocking call.
                      Set to `False` to run in a new subprocess.
@@ -79,6 +82,7 @@ def call(device: str, speed: int, rtos: Path = None, blocking: bool = True,
     binary = os.environ.get("PX4_JLINK_GDB_SERVER", "JLinkGDBServer")
 
     command_jlink = f"{binary} -device {device} -speed {speed} -if swd -noreset -nogui"
+    if serial: command_jlink += f" -SelectEmuBySN {serial}"
     if rtos: command_jlink += f" -rtos {Path(rtos).absolute()}"
     if silent: command_jlink += " -silent"
 
@@ -95,7 +99,7 @@ def call(device: str, speed: int, rtos: Path = None, blocking: bool = True,
 
 
 # -----------------------------------------------------------------------------
-def itm(device: str, baudrate: int = None, port: int = None) -> int:
+def itm(device: str, baudrate: int = None, port: int = None, serial: str = None) -> int:
     """
     Launches `JLinkSWOViewer`, connects to the microcontroller. You can
     overwrite the default binary by exporting an alternative in your environment:
@@ -107,6 +111,7 @@ def itm(device: str, baudrate: int = None, port: int = None) -> int:
     :param device: part name of the microcontroller to debug.
     :param baudrate: optional frequency of the SWO connection.
     :param port: ITM port to display at startup. You can modify this at runtime.
+    :param serial: Serial number of J-Link for selection.
 
     :return: the process return code
     """
@@ -140,7 +145,7 @@ def rtt(backend: JLinkBackend, channel: int = 0) -> int:
     return 0
 
 # -----------------------------------------------------------------------------
-def program(device: str, speed: int, source: Path, load_addr: int = None) -> int:
+def program(source: Path, device: str, speed: int, serial: str = None, load_addr: int = None) -> int:
     """
     Loads the source file into the microcontroller and resets the device. You can
     overwrite the default binary by exporting an alternative in your environment:
@@ -149,9 +154,10 @@ def program(device: str, speed: int, source: Path, load_addr: int = None) -> int
     export PX4_JLINK=path/to/JLinkExe
     ```
 
+    :param source: path to a `.bin`, `.hex`, or `.elf` file to upload.
     :param device: part name of the microcontroller to program.
     :param speed: SWD connection baudrate in kHz.
-    :param source: path to a `.bin`, `.hex`, or `.elf` file to upload.
+    :param serial: Serial number of J-Link for selection.
     :param load_addr: Specifies the start address to load a `.bin` file.
 
     :return: the process return code of JLink
@@ -162,13 +168,15 @@ def program(device: str, speed: int, source: Path, load_addr: int = None) -> int
     with tempfile.NamedTemporaryFile() as fcmd:
         # LoadFile erases the correct amount of sectors, then writes the new
         # binary or hex file. We still need to reset afterwards.
-        Path(fcmd.name).write_text(f"LoadFile {source} {load_addr or ''}\nReset\nExit")
-        jcmd = f"{binary} -device {device} -if swd -speed {speed} " \
-               f"-autoconnect 1 -nogui 1 -commandfile {fcmd.name}"
-        LOGGER.debug(jcmd)
+        commands = f"LoadFile {source} {load_addr or ''}\nReset\nExit"
+        Path(fcmd.name).write_text(commands)
+        jcmd = f"{binary} -device {device} -speed {speed} -ExitOnError 1 " \
+               f"-if swd -autoconnect 1 -nogui 1 -commandfile {fcmd.name}"
+        if serial: jcmd += f" -SelectEmuBySN {serial}"
+        LOGGER.debug(f"{jcmd}\n{commands}")
         return subprocess.call(jcmd, shell=True)
 
-def reset(device: str, speed: int) -> int:
+def reset(device: str, speed: int, serial: str = None) -> int:
     """
     Resets the device via JLink.
 
@@ -179,15 +187,18 @@ def reset(device: str, speed: int) -> int:
                    export PX4_JLINK=path/to/JLinkExe
                    ```
     :param speed: SWD connection baudrate in kHz.
+    :param serial: Serial number of J-Link for selection.
 
     :return: the process return code of JLink
     """
     binary = os.environ.get("PX4_JLINK", "JLinkExe")
     with tempfile.NamedTemporaryFile() as fcmd:
-        Path(fcmd.name).write_text(f"Reset\nExit")
-        jcmd = f"{binary} -device {device} -if swd -speed {speed} " \
-               f"-autoconnect 1 -nogui 1 -commandfile {fcmd.name}"
-        LOGGER.debug(jcmd)
+        commands = "Reset\nExit"
+        Path(fcmd.name).write_text(commands)
+        jcmd = f"{binary} -device {device} -speed {speed} -ExitOnError 1 " \
+               f"-if swd -autoconnect 1 -nogui 1 -commandfile {fcmd.name}"
+        if serial: jcmd += f" -SelectEmuBySN {serial}"
+        LOGGER.debug(f"{jcmd}\n{commands}")
         return subprocess.call(jcmd, shell=True)
 
 
@@ -203,7 +214,11 @@ def _add_subparser(subparser):
             type=int,
             default=12000,
             help="SWO baudrate in kHz.")
-    parser.set_defaults(backend=lambda args: JLinkBackend(args.device, args.speed))
+    parser.add_argument(
+            "-serial",
+            default=None,
+            help="Serial of the J-Link.")
+    parser.set_defaults(backend=lambda args: JLinkBackend(args.device, args.speed, args.serial))
     return parser
 
 
@@ -223,6 +238,10 @@ if __name__ == "__main__":
         type=int,
         default=12000,
         help="SWD baudrate in kHz.")
+    parser.add_argument(
+        "-serial",
+        default=None,
+        help="Serial of the J-Link.")
     parser.add_argument(
         "-v",
         dest="verbosity",
@@ -265,20 +284,20 @@ if __name__ == "__main__":
     emdbg.logger.configure(args.verbosity)
 
     if args.command == "reset":
-        exit(reset(args.device, args.speed))
+        exit(reset(args.device, args.speed, args.serial))
 
     if args.command == "run":
-        exit(call(args.device, args.speed, blocking=True))
+        exit(call(args.device, args.speed, args.serial, blocking=True))
 
     if args.command == "upload":
-        exit(program(args.device, args.speed, os.path.abspath(args.source)))
+        exit(program(os.path.abspath(args.source), args.device, args.speed, args.serial))
 
     if args.command == "rtt":
-        backend = JLinkBackend(args.device, args.speed)
+        backend = JLinkBackend(args.device, args.speed, args.serial)
         exit(rtt(backend, args.channel))
 
     if args.command == "itm":
-        exit(itm(args.device, args.baudrate, args.channel))
+        exit(itm(args.device, args.baudrate, args.channel, args.serial))
 
     LOGGER.error("Unknown command!")
     exit(1)